ml_server/requirements_runpod.txt
# RunPod Serverless handler — torch/torchvision already in base image

runpod>=1.6.0
httpx==0.27.2
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# ML deps
fvcore
cloudpickle
omegaconf
hydra-core
pycocotools
timm
opencv-python-headless
rasterio
numpy
pandas
geopandas
shapely
pyproj
simplekml
